refactor: add MobileProfileDrawer to FriendsPage and SwipesPage
